Refactoring. Extract cluster interactions to separete entity
